create store Modules but need to solve errors.
