+++ src/views/HouseDetail.vue
@@ -73,13 +73,13 @@ export default {
       return parseInt(this.$route.params.id)
     },
     house() {
-      const houses = this.$store.state.houses
+      const houses = this.$store.state.housesModule.houses
       const ahouse = houses.find(house => house.id === this.houseId)
       return ahouse ? ahouse : null
     }
   },
   created() {
-    this.$store.dispatch('fetchHouses');
+    this.$store.dispatch('housesModule/fetchHouses');
   },
 }
 </script>
